added purchase functionality and event retrieval to landing page and buyer dashboard

+++ funflare/src/main/java/com/funflare/funflare/dto/WalletResponseDTO.java
@@ -11,7 +11,7 @@ public class WalletResponseDTO  {
 
     private Long userId;
 
-    private BigDecimal balance;
+    private Double balance;
 
     private OffsetDateTime createdAt;
 
@@ -46,11 +46,11 @@ public class WalletResponseDTO  {
         this.userId = userId;
     }
 
-    public BigDecimal getBalance() {
+    public Double getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
